Add sql scripts, add first selection

+++ src/main/java/com/aneirine/demo/models/User.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class User {
 
-    private long id;
+    private int id;
     private String name;
-    private double salary;
+    private int salary;
 }
